+++ src/commands/MODEL/modelCommandParser.ts
@@ -134,15 +134,15 @@ export const modelCommandParser: PipelineBothCommandParser<ModelCommand> = {
         if ($pipelineJson.defaultModelRequirements[command.key] !== undefined) {
             if ($pipelineJson.defaultModelRequirements[command.key] === command.value) {
                 console.warn(`Multiple commands \`MODEL ${command.key} ${command.value}\` in the pipeline head`);
-                // <- TODO: [🚎] Some better way how to get warnings from pipeline parsing / logic
+                // <- TODO: [🚎][💩] Some better way how to get warnings from pipeline parsing / logic
             } else {
                 throw new ParseError(
                     spaceTrim(`
-                        Redefinition of MODEL \`${command.key}\` in the pipeline head
+                        Redefinition of \`MODEL ${command.key}\` in the pipeline head
 
                         You have used:
-                        - MODEL ${command.key} ${$pipelineJson.defaultModelRequirements[command.key]}
-                        - MODEL ${command.key} ${command.value}
+                        1) \`MODEL ${command.key} ${$pipelineJson.defaultModelRequirements[command.key]}\`
+                        2) \`MODEL ${command.key} ${command.value}\`
                     `),
                 );
             }