+++ src/conversion/prettify/renderPromptbookMermaid.ts
@@ -4,7 +4,8 @@ import { RESERVED_PARAMETER_NAMES } from '../../config';
 import { UnexpectedError } from '../../errors/UnexpectedError';
 import type { PipelineJson } from '../../pipeline/PipelineJson/PipelineJson';
 import type { TaskJson } from '../../pipeline/PipelineJson/TaskJson';
-import type { string_href, string_name } from '../../types/typeAliases';
+import type { string_href } from '../../types/typeAliases';
+import type { string_name } from '../../types/typeAliases';
 import { normalizeTo_camelCase } from '../../utils/normalization/normalizeTo_camelCase';
 import { titleToName } from '../../utils/normalization/titleToName';