+++ src/conversion/utils/extractVariablesFromScript.ts
@@ -1,6 +1,7 @@
 import { spaceTrim } from 'spacetrim';
 import { ParseError } from '../../errors/ParseError';
-import type { string_javascript, string_javascript_name } from '../../types/typeAliases';
+import type { string_javascript } from '../../types/typeAliases';
+import type { string_javascript_name } from '../../types/typeAliases';
 /**
  * Parses the given script and returns the list of all used variables that are not defined in the script
  *
@@ -12,6 +13,7 @@ import type { string_javascript, string_javascript_name } from '../../types/type
 export function extractVariablesFromScript(script: string_javascript): Set<string_javascript_name> {
     const variables = new Set<string_javascript_name>();
 
+    const originalScript = script;
     script = `(()=>{${script}})()`;
 
     try {
@@ -22,21 +24,35 @@ export function extractVariablesFromScript(script: string_javascript): Set<strin
                 if (!(error instanceof ReferenceError)) {
                     throw error;
                 }
-                const undefinedName = error.message.split(' ')[0];
+
                 /*
                 Note: Parsing the error
+                      🌟 Most devices:
                       [PipelineUrlError: thing is not defined]
+
+                      🍏 iPhone`s Safari:
+                      [PipelineUrlError: Can't find variable: thing]
                 */
 
-                if (!undefinedName) {
+                let variableName: string | undefined = undefined;
+
+                if (error.message.startsWith(`Can't`)) {
+                    // 🍏 Case
+                    variableName = error.message.split(' ').pop();
+                } else {
+                    // 🌟 Case
+                    variableName = error.message.split(' ').shift();
+                }
+
+                if (variableName === undefined) {
                     throw error;
                 }
 
-                if (script.includes(undefinedName + '(')) {
-                    script = `const ${undefinedName} = ()=>'';` + script;
+                if (script.includes(variableName + '(')) {
+                    script = `const ${variableName} = ()=>'';` + script;
                 } else {
-                    variables.add(undefinedName);
-                    script = `const ${undefinedName} = '';` + script;
+                    variables.add(variableName);
+                    script = `const ${variableName} = '';` + script;
                 }
             }
     } catch (error) {
@@ -50,6 +66,20 @@ export function extractVariablesFromScript(script: string_javascript): Set<strin
                     Can not extract variables from the script
 
                     ${block((error as Error).toString())}}
+
+
+                    Found variables:
+
+                    ${Array.from(variables)
+                        .map((variableName, i) => `${i + 1}) ${variableName}`)
+                        .join('\n')}
+
+
+                    The script:
+
+                    \`\`\`javascript
+                    ${block(originalScript)}
+                    \`\`\`
                 `,
                 // <- TODO: [🚞] Pass from consumer(s) of `extractVariablesFromScript`
             ),