+++ src/execution/createPipelineExecutor/getExamplesForTask.ts
@@ -13,7 +13,7 @@ import { TODO_USE } from '../../utils/organization/TODO_USE';
 export async function getExamplesForTask(
     task: ReadonlyDeep<TaskJson>,
 ): Promise<string_parameter_value & string_markdown> {
-    // TODO: [♨] Implement Better - use real index and keyword search
+    // TODO: [♨][💩] Implement Better - use real index and keyword search
 
     TODO_USE(task);
     return RESERVED_PARAMETER_MISSING_VALUE /* <- TODO: [♨] Implement */;