+++ src/utils/serialization/$deepFreeze.ts
@@ -1,5 +1,6 @@
 import type { ReadonlyDeep } from 'type-fest';
 import type { really_any } from '../organization/really_any';
+import type { TODO_any } from '../organization/TODO_any';
 
 /**
  * @@@
@@ -11,6 +12,10 @@ import type { really_any } from '../organization/really_any';
  * @public exported from `@promptbook/utils`
  */
 export function $deepFreeze<TObject>(objectValue: TObject): ReadonlyDeep<TObject> {
+    if (Array.isArray(objectValue)) {
+        return Object.freeze(objectValue.map((item) => $deepFreeze(item))) as TODO_any;
+    }
+
     const propertyNames = Object.getOwnPropertyNames(objectValue);
     for (const propertyName of propertyNames) {
         const value = (objectValue as really_any)[propertyName];